adverseiq-backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
asyncpg==0.29.0
httpx==0.27.0
openai==1.30.0
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.2.1
alembic==1.13.1
psycopg2-binary==2.9.11
fpdf2==2.8.3
jinja2==3.1.4
alembic==1.13.1